backend/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
av==15.1.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
ctranslate2==4.6.0
distro==1.9.0
fastapi==0.116.1
faster-whisper==1.2.0
filelock==3.19.1
flatbuffers==25.2.10
fsspec==2025.7.0
groq==0.31.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
humanfriendly==10.0
idna==3.10
inflect==7.5.0
jiter==0.10.0
llvmlite==0.44.0
more-itertools==10.8.0
mpmath==1.3.0
numba==0.61.2
numpy==2.2.6
onnxruntime==1.22.1
openai==1.102.0
packaging==25.0
protobuf==6.32.0
PyAudio==0.2.14
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.5
resampy==0.4.3
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.2
soundfile==0.13.1
srt==3.5.3
starlette==0.47.3
sympy==1.14.0
tokenizers==0.22.0
tqdm==4.67.1
typeguard==4.4.4
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
vosk==0.3.45
websockets==15.0.1
